image-augmentation-aws/lambda-function/rotation-worker/requirements.txt
boto3==1.28.57
botocore==1.31.57
Pillow==10.0.0